+++ src/main/java/com/rednavis/vaadin/showcase/backend/Employee.java
@@ -1,35 +1,17 @@
 package com.rednavis.vaadin.showcase.backend;
 
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
+import java.time.LocalDate;
+import lombok.Data;
 
 /**
  * Model.
  */
-@Getter
-@Setter
-@ToString
+@Data
 public class Employee {
 
-  private String firstname;
-  private String lastname;
+  private String firstName;
+  private String lastName;
   private String title;
   private String email;
-  private String notes = "";
-
-  /**
-   * Employee.
-   *
-   * @param firstname firstname
-   * @param lastname  lastname
-   * @param email     email
-   * @param title     title
-   */
-  public Employee(String firstname, String lastname, String email, String title) {
-    this.firstname = firstname;
-    this.lastname = lastname;
-    this.email = email;
-    this.title = title;
-  }
+  private LocalDate birthDay;
 }
