+++ src/main/java/com/rednavis/vaadin/showcase/backend/service/password/PasswordUtils.java
@@ -12,10 +12,12 @@ import lombok.experimental.UtilityClass;
 class PasswordUtils {
 
   private static final String PBKDF2_ALGORITHM = "PBKDF2WithHmacSHA1";
+
   // These constants may be changed without breaking existing hashes.
   private static final int SALT_BYTE_SIZE = 24;
   private static final int HASH_BYTE_SIZE = 18;
   private static final int PBKDF2_ITERATIONS = 64000;
+
   // These constants define the encoding and may not be changed.
   private static final int HASH_SECTIONS = 5;
   private static final int HASH_ALGORITHM_INDEX = 0;
