--- src/test/java/com/rednavis/vaadin/showcase/TestUtils.java
@@ -1,6 +1,0 @@
-package com.rednavis.vaadin.showcase;
-
-public class TestUtils {
-
-  public static final String PASSWORD = "1@QWaszx";
-}
